COMMERCE-5169 commerce-machine-learning-test Move comparator to the right package
