+++ modules/dxp/apps/commerce/commerce-machine-learning-test/src/testIntegration/java/com/liferay/commerce/machine/learning/recommendation/test/util/comparator/FrequentPatternCommerceMLRecommendationComparator.java
@@ -12,7 +12,7 @@
  *
  */
 
-package com.liferay.commerce.machine.learning.recommendation.test.util;
+package com.liferay.commerce.machine.learning.recommendation.test.util.comparator;
 
 import com.liferay.commerce.machine.learning.recommendation.FrequentPatternCommerceMLRecommendation;
 import com.liferay.portal.kernel.util.SetUtil;
